% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_Sequence_Motif.R
\name{DNA_Sequence_Motif}
\alias{DNA_Sequence_Motif}
\title{Búsqueda de motivos dentro de una secuencia de DNA dada}
\usage{
DNA_Sequence_Motif(Sec, Motif)
}
\arguments{
\item{Sec}{Cadena de caracteres de DNA en la cual se buscará un motivo dado}

\item{Motif}{Cadena de caracter más corto que la secuencia que se buscará en una secuencia dada}
}
\value{
msg
}
\description{
Búsqueda de motivos dentro de una secuencia de DNA dada
}
\examples{
DNA_Sequence_Motif("GTACGTACGTAGCTGACTACGATCAGCGGAGTCGATCATCACGCTACGTACGTACGATGGGATTTTCAGCATGC","gatc")
DNA_Sequence_Motif("GTACGTACGTAGCTGACTACGATCAGCGGAGTCGATCATCACGCTACGTACGTACGATGGGATTTTCAGCATGC","GATA")
DNA_Sequence_Motif("GTACGTACGTAGCTGACTACGATCAGCGGAGTCGATCATCACGCTACGTACGTACGATGGGATTTTCAGCATGC","gatU")
}
